feat: add deletion, revamp 'what's this' section and handle search blur
