+++ app/(main-layout)/about/page.tsx
@@ -1,24 +1,26 @@
-import Image from 'next/image';
 import styles from './page.module.css';
-import Link from 'next/link';
 
 const About = () => {
   return (
-    <div>
-      <section>
-        <h2 className="test">All your beloved trips in one place</h2>
-        <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Doloremque, ipsa accusantium ratione voluptate est deleniti reprehenderit tempora cumque praesentium esse in! Quia vel nisi molestiae facere laudantium rerum placeat laboriosam?</p>
+    <div className={styles.container}>
+      <h1>What's this?</h1>
+      
+      <section className={styles.section}>
+        <h2>All your beloved trips in one place</h2>
+        <p>Plan and organize your travel destinations effortlessly. Save places you want to visit and schedule them for future trips. Never lose track of your travel wishlist again!</p>
       </section>
-      <section>
-        <h2>Powered with _some_ AI</h2>
-        <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Doloremque, ipsa accusantium ratione voluptate est deleniti reprehenderit tempora cumque praesentium esse in! Quia vel nisi molestiae facere laudantium rerum placeat laboriosam?</p>
+
+      <section className={styles.section}>
+        <h2>Powered with Google Places API</h2>
+        <p>Search for any place worldwide using Google's extensive database. Get accurate location suggestions and save them to your personal collection with just a click.</p>
       </section>
-      <section>
+
+      <section className={styles.section}>
         <h2>Calendar connectivity on sight!</h2>
-        <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Doloremque, ipsa accusantium ratione voluptate est deleniti reprehenderit tempora cumque praesentium esse in! Quia vel nisi molestiae facere laudantium rerum placeat laboriosam?</p>
+        <p>Organize your visits by date, group them automatically, and keep track of your travel schedule. Perfect for planning short trips or extended adventures.</p>
       </section>
     </div>
-  )
-}
+  );
+};
 
 export default About;
